+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changes
 
+- `openModal` and `openWindow` returns the Stage
+- `dialog` builder operates on a StageAwareFieldset so it can close the dialog easier by calling `close()`
 - All JSON extractor functions support vararg keys, will pick the first available (https://github.com/edvin/tornadofx/issues/350)
 
 ## [1.7.7] - 2017-06-15